test: extract gateway subscription parameter fixture
